fix: split lib code for server-side rendering
