+++ apps/main/src/loan/components/PageLlamaMarkets/Page.tsx
@@ -9,9 +9,9 @@ import { setSupportedChains, setSupportedLendingChains } from '@/loan/entities/c
 import {
   invalidateAllUserLendingVaults,
   invalidateLendingVaults,
-  invalidateAllUserLendingSupplies,
   setLendingVaults,
 } from '@/loan/entities/lending-vaults'
+import { invalidateAllUserLendingSupplies } from '@/loan/entities/lending-vaults-rpc'
 import { useLlamaMarkets } from '@/loan/entities/llama-markets'
 import { invalidateAllUserMintMarkets, invalidateMintMarkets, setMintMarkets } from '@/loan/entities/mint-markets'
 import useStore from '@/loan/store/useStore'
