+++ apps/main/src/loan/entities/lending-vaults.ts
@@ -1,11 +1,8 @@
-import { enforce, group, test } from 'vest'
 import {
   type UserContractParams,
   type UserContractQuery,
   userContractValidationSuite,
 } from '@/loan/entities/user-contract'
-import { ChainId } from '@/loan/types/loan.types'
-import type { LlamaApi } from '@/loan/types/loan.types'
 import { Chain as ChainName } from '@curvefi/prices-api'
 import {
   getAllMarkets,
@@ -16,13 +13,10 @@ import {
   type UserMarketEarnings,
   type UserMarketStats,
 } from '@curvefi/prices-api/llamalend'
-import { getLib, requireLib } from '@ui-kit/features/connect-wallet'
-import { type ChainParams, type ChainQuery, queryFactory, UserParams, type UserQuery } from '@ui-kit/lib/model/query'
-import { apiValidationGroup, chainValidationGroup } from '@ui-kit/lib/model/query/chain-validation'
-import { userAddressValidationGroup, userAddressValidationSuite } from '@ui-kit/lib/model/query/user-address-validation'
-import { createValidationSuite, EmptyValidationSuite } from '@ui-kit/lib/validation'
+import { queryFactory, UserParams, type UserQuery } from '@ui-kit/lib/model/query'
+import { userAddressValidationSuite } from '@ui-kit/lib/model/query/user-address-validation'
+import { EmptyValidationSuite } from '@ui-kit/lib/validation'
 import { type Address } from '@ui-kit/utils'
-import networks from '../networks'
 
 export type LendingVault = Market & { chain: ChainName }
 
@@ -56,44 +50,6 @@ const {
   validationSuite: userAddressValidationSuite,
 })
 
-const {
-  getQueryOptions: getUserLendingSuppliesQueryOptions,
-  getQueryData: getCurrentUserLendingSupplies,
-  invalidate: invalidateUserLendingSupplies,
-} = queryFactory({
-  queryKey: ({ chainId, userAddress }: UserParams & ChainParams<ChainId>) =>
-    ['user-lending-supplies', { chainId }, { userAddress }, 'v2'] as const,
-  queryFn: async (_: UserQuery & ChainQuery<ChainId>) => {
-    const api = requireLib<LlamaApi>()
-    await api.lendMarkets.fetchMarkets()
-    return {
-      // todo: multi-chain not supported
-      [networks[api.chainId].name]: Object.fromEntries(
-        await Promise.all(
-          api.lendMarkets.getMarketList().map(async (name) => {
-            const { addresses, vault, wallet } = api.getLendMarket(name)
-            const { vaultShares } = await wallet.balances()
-            return [addresses.controller, +vaultShares && +(await vault.convertToAssets(vaultShares))]
-          }),
-        ),
-      ),
-    } as Record<ChainName, Record<Address, number>>
-  },
-  validationSuite: createValidationSuite(({ chainId, userAddress }: UserParams & ChainParams<ChainId>) => {
-    userAddressValidationGroup({ userAddress })
-    chainValidationGroup({ chainId })
-    apiValidationGroup({ chainId })
-    group('userAddressLibCheck', () => {
-      test('userAddress', 'Invalid EVM address', () => {
-        const lib = getLib<LlamaApi>()
-        if (userAddress && lib) {
-          enforce(userAddress).equals(lib.signerAddress)
-        }
-      })
-    })
-  }),
-})
-
 const { useQuery: useUserLendingVaultStatsQuery, invalidate: invalidateUserLendingVaultStats } = queryFactory({
   queryKey: ({ userAddress, contractAddress, blockchainId }: UserContractParams) =>
     ['user-lending-vault', 'stats', { blockchainId }, { contractAddress }, { userAddress }, 'v1'] as const,
@@ -102,7 +58,7 @@ const { useQuery: useUserLendingVaultStatsQuery, invalidate: invalidateUserLendi
   validationSuite: userContractValidationSuite,
 })
 
-const { useQuery: useUserLendingVaultEarningsQuery, invalidate: invalidateUserLendingVaultEarnings } = queryFactory({
+export const { useQuery: useUserLendingVaultEarnings, invalidate: invalidateUserLendingVaultEarnings } = queryFactory({
   queryKey: ({ userAddress, contractAddress, blockchainId }: UserContractParams) =>
     ['user-lending-vault', 'earnings', { blockchainId }, { contractAddress }, { userAddress }, 'v1'] as const,
   queryFn: async ({ userAddress, contractAddress, blockchainId }: UserContractQuery): Promise<UserMarketEarnings> =>
@@ -120,17 +76,5 @@ export function invalidateAllUserLendingVaults(userAddress: Address | undefined)
   })
 }
 
-export function invalidateAllUserLendingSupplies(userAddress: Address | undefined) {
-  Object.entries(getCurrentUserLendingSupplies({ userAddress }) ?? {}).forEach(([chain, positions]) => {
-    invalidateUserLendingSupplies({ userAddress })
-    const blockchainId = chain as ChainName
-    ;(Object.keys(positions) as Address[]).forEach((contractAddress) => {
-      invalidateUserLendingVaultEarnings({ userAddress, blockchainId, contractAddress })
-    })
-  })
-}
-
 export const getUserLendingVaultsOptions = getUserLendingVaultsQueryOptions
-export const getUserLendingSuppliesOptions = getUserLendingSuppliesQueryOptions
 export const useUserLendingVaultStats = useUserLendingVaultStatsQuery
-export const useUserLendingVaultEarnings = useUserLendingVaultEarningsQuery
